+++ Chapter.Net.WPF.Converters/IntegerToBooleanConverter/IntegerToBooleanConverter.cs
@@ -5,7 +5,9 @@
 // -----------------------------------------------------------------------------------------------------------------
 
 using System;
+using System.ComponentModel;
 using System.Globalization;
+using System.Linq;
 using System.Windows.Data;
 
 // ReSharper disable once CheckNamespace
@@ -13,20 +15,28 @@ using System.Windows.Data;
 namespace Chapter.Net.WPF.Converters;
 
 /// <summary>
-///     Converts an integer to its boolean representation and back.
+///     Converts a single integer or a list of integers to a boolean representation.
 /// </summary>
 [ValueConversion(typeof(int), typeof(bool))]
+[ValueConversion(typeof(int[]), typeof(bool))]
 public class IntegerToBooleanConverter : SingleAndMultiValueConverter
 {
     /// <summary>
-    ///     Converts an integer to its boolean representation.
+    ///     Defines the boolean to use if the integers are mixed.
     /// </summary>
-    /// <param name="value">The integer to convert.</param>
+    /// <value>Default: false.</value>
+    [DefaultValue(false)]
+    public bool? MixedIs { get; set; } = false;
+
+    /// <summary>
+    ///     Converts a single integer to a boolean representation.
+    /// </summary>
+    /// <param name="value">The value to convert.</param>
     /// <param name="targetType">Unused.</param>
     /// <param name="parameter">Unused.</param>
     /// <param name="culture">Unused.</param>
-    /// <returns>The converted boolean value.</returns>
-    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    /// <returns>The converted value.</returns>
+    public override object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
         if (value is int integer)
             return System.Convert.ToBoolean(integer);
@@ -34,17 +44,37 @@ public class IntegerToBooleanConverter : SingleAndMultiValueConverter
     }
 
     /// <summary>
-    ///     Converts a boolean to its integer representation.
+    ///     Converts a list of integers to a boolean representation.
     /// </summary>
-    /// <param name="value">The boolean to convert.</param>
+    /// <param name="values">The values to convert.</param>
     /// <param name="targetType">Unused.</param>
     /// <param name="parameter">Unused.</param>
     /// <param name="culture">Unused.</param>
-    /// <returns>The converted integer value.</returns>
-    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    /// <returns>The converted value.</returns>
+    public override object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
     {
-        if (value is bool boolean)
-            return System.Convert.ToInt32(boolean);
-        return System.Convert.ToInt32(true);
+        if (values == null)
+            return false;
+
+        var booleans = values.OfType<int>().Distinct().ToList();
+        return booleans.Count switch
+        {
+            0 => false,
+            > 1 => MixedIs,
+            _ => System.Convert.ToBoolean(booleans[0])
+        };
+    }
+
+    /// <summary>
+    ///     Converts a single boolean back to it integer representation.
+    /// </summary>
+    /// <param name="value">The value to convert.</param>
+    /// <param name="targetType">Unused.</param>
+    /// <param name="parameter">Unused.</param>
+    /// <param name="culture">Unused.</param>
+    /// <returns>The converted value.</returns>
+    public override object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        return value is bool boolean ? System.Convert.ToInt32(boolean) : System.Convert.ToInt32(true);
     }
 }